# vectorized numpy version