I've been using your NPEET package to estimate mutual information between hidden layers of neural nets and decided to vectorize its implementation to match numpy style. Consider merging.
Here is the old versus new timings.

```python
import numpy as np
import npeet.entropy_estimators as ee  # current list version
import npeet.npeet as ee2  # vectorized numpy version

n = 1000


def to_list(x):
    return list(map(list, x))


def to_tuple(x):
    return list(map(tuple, x))


def setup_x():
    return np.random.rand(n, 10)


def setup_y():
    return np.random.rand(n, 3)


def setup_z():
    return np.random.rand(n, 5)


print('mi(x, y):')
%timeit ee.mi(setup_x().tolist(), setup_y().tolist())
%timeit ee2.mi(setup_x(), setup_y())

print('ctc(x, y):')
%timeit ee.ctc(setup_x().tolist(), setup_y().tolist())
%timeit ee2.ctc(setup_x(), setup_y())

print('cmi(x, y, z):')
%timeit ee.cmi(setup_x().tolist(), setup_y().tolist(), z=setup_z().tolist())
%timeit ee2.mi(setup_x(), setup_y(), z=setup_z())

print('micd(x, y):')
%timeit ee.micd(setup_x().tolist(), np.random.randint(0, 10, (n, 3)).tolist(), warning=False)
%timeit ee2.micd(setup_x(), np.random.randint(0, 10, (n, 3)), warning=False)


for i in range(10):
    x, y = setup_x(), setup_y()
    x_other = setup_x()
    z = setup_z()
    x_list = to_list(x)
    y_list = to_list(y)

    x_discrete = np.random.randint(0, 10, size=x.shape)
    y_discrete = np.random.randint(0, 10, size=y.shape)
    z_discrete = np.random.randint(0, 10, size=z.shape)

    assert np.isclose(ee2.entropy(x), ee.entropy(x_list))
    assert np.isclose(ee2.centropy(x, y), ee.centropy(x_list, y_list))
    assert np.isclose(ee2.tc(x), ee.tc(x_list))
    assert np.isclose(ee2.ctc(x, y), ee.ctc(x_list, y_list))
    assert np.isclose(ee2.corex(x, y), ee.corex(x_list, y_list))
    assert np.isclose(ee2.mi(x, y), ee.mi(x_list, y_list))
    assert np.isclose(ee2.mi(x, y, z=z), ee.cmi(x_list, y_list, z))
    assert np.isclose(ee2.kldiv(x, x_other), ee.kldiv(x_list, to_list(x_other)))

    assert np.isclose(ee2.midd(x_discrete, y_discrete), ee.midd(to_tuple(x_discrete), to_tuple(y_discrete)))
    assert np.isclose(ee2.cmidd(x_discrete, y_discrete, z_discrete),
                      ee.cmidd(to_tuple(x_discrete), to_tuple(y_discrete), to_tuple(z_discrete)))
    assert np.isclose(ee2.micd(x, y_discrete, warning=False), ee.micd(x_list, y_discrete.tolist(), warning=False))

```

Output
```
mi(x, y):
194 ms ± 1.25 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
83.2 ms ± 2.62 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
ctc(x, y):
1.74 s ± 1.67 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
62.5 ms ± 16.4 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
cmi(x, y, z):
309 ms ± 2.04 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
167 ms ± 1.92 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
micd(x, y):
161 ms ± 727 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
45.1 ms ± 988 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

`avgdigamma` is the most time consuming function but I don't know how to vectorize the internal for-loop.