@gregversteeg 2019-01-18
Thank you so much! When I wrote this, I hadn't absorbed the numpy style. I always meant to go back and vectorize, but haven't had the time. This is being greedy... but could you look at the readme and change the example if you think it can be simplified using your merge?