@dizcza 2019-01-18
Both examples in readme still work, because input lists are cast to numpy arrays.
You can remove the part "Note that we converted the numpy arrays to lists..."